// src/Frontend/src/app/server/model/applicationModelsApiModelsApiTimeSpecificationDate.ts
/**
 * WebApi
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ApplicationModelsApiModelsApiTimeSpecificationMoment } from './applicationModelsApiModelsApiTimeSpecificationMoment';

export interface ApplicationModelsApiModelsApiTimeSpecificationDate extends ApplicationModelsApiModelsApiTimeSpecificationMoment { 
    date?: string;
}