// src/Frontend/src/app/server/model/applicationModelsApiModelsApiTimeSpecificationTypes.ts
/**
 * WebApi
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type ApplicationModelsApiModelsApiTimeSpecificationTypes = 'Base' | 'Period' | 'Moment' | 'Date' | 'DateTime' | 'Month';

export const ApplicationModelsApiModelsApiTimeSpecificationTypes = {
    Base: 'Base' as ApplicationModelsApiModelsApiTimeSpecificationTypes,
    Period: 'Period' as ApplicationModelsApiModelsApiTimeSpecificationTypes,
    Moment: 'Moment' as ApplicationModelsApiModelsApiTimeSpecificationTypes,
    Date: 'Date' as ApplicationModelsApiModelsApiTimeSpecificationTypes,
    DateTime: 'DateTime' as ApplicationModelsApiModelsApiTimeSpecificationTypes,
    Month: 'Month' as ApplicationModelsApiModelsApiTimeSpecificationTypes
};